refactor benchmark files
